Move decorator into errors module.
Fixes according to comments in PR.
